monitor-dns/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
PiHole-api==2.6
psutil==5.9.0
psycopg2-binary==2.9.3
PyHole==1.6
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
python-crontab==2.6.0
pytz